requirements.txt
Django>=5.0,<6.0
django-unfold>=0.40.0
psycopg2-binary>=2.9
django-tailwind[reload]>=3.8
django-allauth[socialaccount]>=65.0
Pillow>=10.0
pytesseract>=0.3
pdfminer.six>=20221105
pdf2image>=1.17
celery[redis]>=5.3
python-decouple>=3.8
whitenoise>=6.6
gunicorn>=22.0
openpyxl>=3.1
django-axes[ipware]>=7.0
